Remove admin delete permission for need_followup
For reviewed TaggedTraits that need followup, remove the delete
permission from the admin interface using TaggedTraitAdmin. This undoes
some of the changes from commit 116933c. This addresses the issue that,
though the delete button on the admin change view properly archives a
need_followup tagged trait, the resulting success message says that the
tagged trait has been "deleted", which could be confusing.