Make creator fields omit unused values
Where 'creator' is a field listed in the admin, have it use the
(new since 1.8) RelatedOnlyFieldListFilter, which means that users who
have not been used as creators for any of the object will not be shown
in the filter list. So only users who have created DCCReviews will be
listed in the filter list there, e.g.