Prevent TaggedTrait editing in Admin
Make tag and trait readonly_fields in the TaggedTraitAdmin, so that
this model can't be edited on the admin interface. Editing should
instead happen through the UpdateView that was carefully designed for
this purpose. Making had_change_permission would prevent admin editing,
but also would remove the TaggedTrait model from displaying in the admin
interface.