Change display of other tags in taggedtrait_info
In the _taggedtrait_info.html panel, move logic for whether or not to
show other tags out of the template and into the views that use this
panel (TaggedTraitDetail, StudyResponseCreateDisagree, DCCReviewCreate,
DCCReviewUpdate, and DCCReviewByTagAndStudy). Require a context variable
('show_other_tags') to turn on display of other tags linked to the same
trait. Also require new context variables, 'other_tags' and
'archived_other_tags', to provide the data to display. Remove use of the
old argument ('show_all_tags') to the included template. Add the new
context variables to the DCCReviewMixin and TaggedTraitDetail views. Add
tests to ensure that the tags are shown or not as appropriate.