Display archived status on DCCReviewNeedFollowupList
Change DCCReviewTable to display archived status for the
DCCReviewNeedFollowupList view. Move status strings into constant
variables in the table class. Change the render_study_response() method
to add a status for archived tagged traits, and also show the study
response status if it is available (there will be two statuses for those
that are responded to by the study and then archived, and only one
status for those that are archived by the DCC without receiving a study
response). Add tests. Change button color for archived tagged traits in
the _taggedtrait_info.html panel to match the contextual text in this
table.