Move detail links column to restricted tables
Move the details column of links to tagged trait detail pages from the
base TaggedTraitTable to the TaggedTraitTableReviewStatusAndDetailsMixin
(preivously the TaggedTraitTableReviewStatusMixin). This means that the
TaggedTraitByTagAndStudyList won't display links to the
TaggedTraitDetail view unless the user is a staff member or tagger for
that study. This accounts for the new permission restrictions on
TaggedTraitDetail.