Add info about deprecation on SourceDatasetDetail page
Add a new context flag that indicates if a deprecation message
should be shown on the SourceDatasetDetail page. Add tests to check
the behavior of this flag. Add some text in the template to check
that is it processing the deprecation message, but the placement
and text themselves will be updated later.