Bulk create traits from the same study version in tests
Change tests that create traits in bulk to create them from the
same study version. When source traits are created in bulk but only
a study is specified, the factories create a different study
version for each trait. Change the tags forms and views tests
related to tagging many traits to create them from the same study
version.