Add tagging form errors and tests for deprecated traits
Force tagging forms to be invalid if a tagged trait is from a
deprecated study version. Add tests to check this behavior. For
most forms, no alteration to the clean method was needed; the
super().clean() method took care of it because the SourceTrait
queryset was already filtered to current() traits in the __init__
method. In those cases, add a comment about why the clean method
does not need changes. The only method that required a change was
the clean method for TagSpecificTraitForm.