Handle two non-deprecated study versions
Handle the existence of two non-deprecated study versions when
checking for the current version of a dataset or trait. Break ties
between two non-deprecated study versions by first selecting the
highest version, followed by the latest date added. This ensures
that only one dataset or trait will be returned, and that it will
have the highest version and date added.