Stop using overly broad permission migrations
Previous permission migrations often added permissions for every model
in a given app to a group or groups. This was problematic because of the
way that Permission objects are created during the migration process.
See here for more information:
    https://code.djangoproject.com/ticket/23422
When Django runs migrations on a brand new database (as it does in tests
or new installations), it first figures out all of the models to make
ContentTypes for. Then it creates the permissions that are used in the
permission data migrations. So the permission data migrations that
added app-wide permissions to a group were also adding permissions for
model types that did not exist at the time the migration was written.
This can lead to unintended permissions for different groups, and
possibly errors in controlling view/data access.

Edit previous permission data migrations that added app-wide permissions
(migrations 2, 3, 4, 6, and 7 of the core app) to fix this issue. Make
sure that models are explicitly named when specifying permissions to
add to a group. Especially problematic was migration 2, which gave
dcc_developers permissions for all models for all apps except for admin,
auth, user, trait_browser, sites, and contenttypes. Now that migration
just explicitly adds permissions for the FlatPage model.

Also remove permissions for the Sessions model from the dcc_developer
group, and remove it from the list of expected permissions in tests.