Add DCCDecision permission migration and improve tests
Add a data migration (0008 for the core app) that adds appropriate
DCCReview permissions to the dcc_developers and dcc_analysts groups.
Improve the tests in core.test_migrations.GroupMigrationTest.
Specifically, remove some repetitive code and fix and indentation error
that led to running the assert statement at the wrong point.