Add basic view for making dcc decisions
Add a view, TaggedTraitsNeedDCCDecisionByTagAndStudyList, which shows a
table of tagged traits that need dcc decisions for a single tag + study
combination. Add a url and template for the view. Add a link to the new
view from the table in TaggedTraitsNeedDCCDecisionSummary. Add a table
class, TaggedTraitDCCDecisionTable, for showing these tagged traits in
the new view.