Add DCCDecision forms and tests
After deciding to implement decision-making loop for DCCDecisions, add
base form and two different forms for creating/updating DCCDecisions in
the loop views and in the create and update views. Also add tests of
these forms. Also move the SubmitCssClass class to the top of the forms
file, since it is used by multiple classes.