Change buttons for dcc decisions
On the need dcc decision summary page, change from a single button that
either says "make" or "update" to one button that says "make final
decisions" (if needed) and one that says "view decisions" (always
there). And in the table shown on the tagged trait need dcc decision
tag and study list page, change from two buttons (one to confirm and one
to remove) to a single button for "make decision".