Automatically archive tagged traits after decision
Modify DCCDecisionMixin so that creating or editing a DCCDecision will
automatically archive a tagged trait with a remove dcc decision and
unarchive a tagged trait with a confirm decision. Add tests of the three
views that create or update dcc decisions. This currently won't work for
changing a dcc decision to confirm from remove, because I realized that
preventing updating an archived tagged trait's dcc decision will mean
that no "remove" dcc decisions can be updated (since they'll be auto-
archived). So I'll need to change that next.