Allow updating archived tagged trait decisions
Modify DCCDecisionUpdate to allow for updating archived tagged traits,
if they have a dcc decision already (if they are archived but missing a
dcc decision, they'll be switched to the create view, which does not
allow creating a decision for an archived tagged trait). Update the
tests to account for new behavior.