Fix skip functionality for DCCDecision loop
Change the DCCDecision.comment field to blank=True, which fixes the skip
functionality in the DCCDecision loop. This wasn't working before
because the comment field was blank=False, which made it into an input
field with the "required" html5 tag, and both Safari and Chrome won't
even submit a post request for a form when a required input field is
empty. So the skip button didn't work, even though the tests were
passing. Making the comment field blank=True fixes this, and I added a
custom clean method to the DCCDecisionBaseForm so that comment can't be
blank.