Remove typo from quality review panel
During the process of testing something on this template I had inserted
the word "COMMENT" where the review comment was supposed to show up, but
when I eventually deleted that I must have overlooked the C because there
was a C showing up at the beginning of every review commen in the quality
review panel. Remove it in this hotfix.