Fix tests of quality review panel
Commit 8a41e424e15958012551883708ffc626b4974f3e, which was performed on
the staging server during the release process, changed the display text
in the quality review panel for tagged traits. This broke tests that
asserted the presence or absence of specific snippets of text from the
panel, especially tests of the TaggedTraitDetail and DCCDecisionCreate
or DCCDecisionUpdate views. Update the snippets of text that are tested
for and make sure the tests pass.