Improve export_tagging to comply with tests
Make lots of small changes that made the tests easier to write. In each
method that creates a file, return the file name rather than None. In
_compress_directory, make variable names more informative. Remove
extraneous spaces from delimiters in the output from the data
dictionaries and the tagged trait data file. Change argument name in
_get_tagged_trait_data for better consistency and clarity. Change
non_archived to include_non_archived and current to include_deprecated,
both with a default of False. In _get_formatted_path_and_make_directory,
use a prefix variable to add 'TOPMed_DCC' to each file name
consistently. Fix a typo where the output directory was missing a /.