Ensure that views in the tags app don't show deprecated traits
Add handling of deprecated study versions to views in the tags app
that didn't already handle them. In general, this involved modifying
the views handling DCCReviews, StudyResponses, and DCCDecisions. Add
tests to check that deprecated tagged traits are properly handled
for all views that didn't already have them. Update one of the
TaggedTraitByTagAndStudyList view and one of the tests for it,which
had a bug that caused deprecated traits to be shown to study
phenotype taggers but not to DCC staff.