Update Study methods for deprecated tagged traits
Update the Study methods that retrieve and count various types of
tagged traits to exclude any tagged trait applied to a deprecated
source trait. Add tests to check specifically for the behavior
with deprecated tagged traits.

Note that the proper way to do this would have been
to move these methods to the SourceStudyVersion model and then
retrieve only the tagged traits attached to that study version, but
that would require so many changes that it will be its own branch.
This is a reasonable interim solution.