Add view showing new variables in most recent study version
Add a view that shows only the new variables in the most recent
study version by checking for accession numbers that only exist in
the this version but not the previous version. Update the
trait_browser urls to add a new url for this view. Add a template
based off of the original study_sourcetrait_list template, with
only a small number of modifications at this point.