Speed up the SourceStudyVersion apply_previous_tags method
Loop only over traits that have a previous version with at least
one tag instead of looping over all traits. This should improve the
speed for studies that have a large number of traits (eg FHS), but
does require an additional database query to obtain the list of
traits in the previous version that have at least one non-archived
tag applied to them, but overall it is likely to be faster.