Test fix
Fix a test of SourceStudyVersion.apply_previous_tags where the
version number of a study version created for the test was not set
properly. This sometimes caused the function that looked up the
previous study version to return an incorrect version, depending on
how the version and timestamp were (randomly) set in the factory.