Filter table queries by db name
LockSourceDbTest and UnlockSourceDbTest were both having issues with
duplicate table names showing up in the query results. I encountered
this problem earlier and thought I'd fixed it but I guess not. A bit of
investigation on the MariaDB server revealed that the query results
sometimes included tables from other devel dbs as well, which resulted
in duplicates. Adding an argument to filter to just the current db
fixed this. This is a special argument for the SHOW TABLES command,
SHOW TABLES FROM db_name.