Get BackupTest working again, finally
Make BackupTest inherit from TransactionTestCase instead of TestCase.
This means that the database is "rolled back" at the end of each test by
truncating each table, rather than by using database transactions that
are literally rolled back at the end of each test. This prevents the
error that I had been getting in this test for so long:
    TransactionManagementError ("An error occurred in the current
    transaction. You can't execute queries until the end of the 'atomic'
    block.")
I think the issue is that the use of dbbackup screwed up the transaction
somehow, but I'm still not exactly sure what was going on. But these
changes fix it.