Redesign TaggedTraitDetail context for deprecation
Replace 'deprecation_message' and 'show_deprecation_message' context
variables with 'is_deprecated', 'show_removed_text', and
'new_version_link'. This keeps all of the queries and model method calls
in the view, but makes it easier to test the context in a stable way.
Update tests to just test the context variables, rather than relying on
matching some text that might be changed later on. Update the template
to use the new context variables appropriately.