Improve TaggedTraitDetailTestsMixin test objects
In TaggedTraitDetailTestsMixin methods that test the deprecation/removal
messages for deprecated tagged traits, explicitly create every version
of every source study version, source dataset, and source trait that are
needed for testing. This makes it easier to see exactly what's going on,
ensures testing conditions are closer to the state of the production
database, and makes it easier to copy/paste this code into other tests.