Update deprecation display for SourceTraitDetail
Similar to the changes in the previous three commits, update the way
that the SourceTraitDetail view displays information about whether a
source trait has been deprecated or removed in a newer study version.
Set new context variables in the view. Use if statements in the template
to show the message as appropriate. Update the tests for the new
changes.