Update deprecation display on SourceDatasetDetail
Update the way the SourceDatasetDetail view uses context variables to
determine whether a dataset has been removed or updated in a newer
study version. Update the way the template uses those context variables
to show a message about the newer version or removal of the dataset.
Update the tests for the new setup, and add a test for when there are
two newer versions.