Apply tags to updated traits in import_db
Add functionality to import_db to apply tags from older trait versions
to the new versions that have just been imported. Add the new method
_apply_tags_to_new_sourcestudyversions. Call the new method within
_import_source_tables. Add a command line argument taggedtrait_creator
and pass it through the various function calls in order to set the
user who is the creator of the updated tagged traits that are created.