Test for incomplete reviews in import_db
In import_db's _import_source_tables, put the call to
_apply_tags_to_new_sourcestudyversions() into an if/else block that
tests to see if there are any incomplete taggedtrait reviews. If any are
found, skip them and print a list of the study version pks to use later
to redo the applying updated tags.

Also update test_import_db for recent changes. Add the
--taggedtrait_creator argument to all calls of call_command for the
import_db management command. Also create a user to use in this argument
where necessary. Add confirmed DCCReviews for all of the taggedtraits
created in ApplyTagsToNewSourceStudyVersionsTest tests to prevent the
new errors in apply_previous_tags() methods.