Fix BackupTest from test_import_db
Move the creation of a test user from .setUpClass() to .setUp(), to fix
the BackupTest class. As it was, test_backup_is_created() would run
second (alphabetically) and result in an error message that the user
did not exist. This is because TransactionTestCase truncates all tables
from the Django test database at the end of every test method.

This issue happened because devel (containing the changes from
feature/move-to-uwit) contained a new test method in BackupTest.
Meanwhile, feature/apply-tags-to-updated-trait-versions added the new
taggedtrait_creator argument, which necessitated adding the fake user in
order to run management.call_command('import_db'), but only had one test
method in BackupTest so that adding the user in the .setUpClass() didn't
cause an error.