Fix import_db test for skipping applying updated tags
Fix a test in import_db's _import_source_tables(). The test is supposed
to determine if there is any unfinished reviewing to be done, but the
logic wasn't quite right. I tried to use the TaggedTraits.need_decision()
queryset to determine if there were unfinished reviews, but the docstring
for need_decision() and need_study_response() are very clear that these
querysets include those that DID NEED a decision or response but have
since had them completed. So I rewrote the logic in _import_source_tables()
to do some manual filtering for three specific scenarios:
    1. no dcc review, dcc decision, or study response
    2. followup, no response, no decision
    3. followup, disagree response, no decision
_apply_tags_to_new_sourcestudyversions() is only run if there are no
taggedtraits of the above three types. Otherwise it's skipped with a
warning message.
Also modify the tests in test_import_db.IntegrationTest to make sure
that _apply_tags_to_new_sourcestudyversions() runs when it is expected
to and does not run when it is expected not to.