Fix SourceDataset.dbgap_link setting
Commit adc811 broke the dbgap dataset links. The end of the URL for a
dataset page should have been "pht=5", but instead it was set to
"pht=pht000005.v1.p1". Replace the use of full_accession in the
setting of this link field with i_accession. I determined that this bug
was never deployed onto the production server.

See Asana task for more details:
https://app.asana.com/0/172540198782525/1128439775764428/f